+++ Anyaroth/Anyaroth/TextureComponent.h
@@ -43,7 +43,7 @@ class TextureComponent
 
 		void load(string filename, uint numRows = 1, uint numCols = 1);
 		void render(const SDL_Rect& rect, double angle = 0, SDL_Point anchor = {0,0}, SDL_RendererFlip flip = SDL_FLIP_NONE) const;
-		void renderFrame(const SDL_Rect& destRect, int row, int col, int angle = 0, SDL_Point anchor = {0,0}, SDL_RendererFlip flip = SDL_FLIP_NONE) const;
+		void renderFrame(const SDL_Rect& destRect, int row, int col, double angle = 0, SDL_Point anchor = {0,0}, SDL_RendererFlip flip = SDL_FLIP_NONE) const;
 		void loadFromText(string text, const Font* font, SDL_Color color);
 
 };
